Revert "Iss134: Implement basic geometry API"
