--- Configuration/python/det_id_check.py
@@ -1,10 +1,0 @@
-import sys
-
-from LDMX.Framework import ldmxcfg
-
-p=ldmxcfg.Process("test")
-p.libraries.append("libEventProc.so")
-detIDChecker = ldmxcfg.Producer("ecalDigis","ldmx::DetIDAnalyzer")
-p.sequence=[detIDChecker]
-p.inputFiles=["input.root"]
-p.printMe()
