+++ DetDescr/CMakeLists.txt
@@ -2,5 +2,4 @@
 module(
   NAME DetDescr
   EXTERNAL_DEPENDENCIES ROOT
-  EXECUTABLES src/import_detector.cxx
 )
