+++ DetDescr/include/DetDescr/HcalID.h
@@ -1,6 +1,6 @@
 /**
  * @file HcalID.h
- * @brief Class that defines an HCal detector ID
+ * @brief Class that defines an HCal sensitive detector
  * @author Jeremy McCormick, SLAC National Accelerator Laboratory
  */
