--- DetDescr/src/DetectorElementImpl.cxx
@@ -1,3 +1,0 @@
-#include "DetDescr/DetectorElementImpl.h"
-
-#include "DetDescr/DetectorDataService.h"
