+++ DetDescr/src/DetectorID.cxx
@@ -2,17 +2,6 @@
 
 namespace ldmx {
 
-    std::ostream& operator<<(std::ostream& stream, FieldValueList& id) {
-        stream << "[";
-        for (int iField = 0; iField < id.size(); iField++) {
-            stream << id[iField];
-            if (iField != id.size() - 1)
-                stream << "/";
-        }
-        stream << "]";
-        return stream;
-    }
-
     DetectorID::~DetectorID() {
         // Delete the field list and its objects as we assume this class owns it.
         for (IDField::IDFieldList::iterator it = fieldList_->begin(); it != fieldList_->end(); it++) {
@@ -37,7 +26,7 @@ namespace ldmx {
         this->rawValue_ = rawValue;
     }
 
-    const FieldValueList& DetectorID::unpack() {
+    const DetectorID::FieldValueList& DetectorID::unpack() {
         std::fill(fieldValues_.begin(), fieldValues_.end(), 0);
         for (IDField::IDFieldList::iterator it = fieldList_->begin();
                 it != fieldList_->end(); it++) {
@@ -105,8 +94,4 @@ namespace ldmx {
         this->fieldValues_.resize(fieldList_->size());
     }
 
-    void DetectorID::clear() {
-        std::fill(fieldValues_.begin(), fieldValues_.end(), 0);
-        rawValue_ = 0;
-    }
 }
