--- DetDescr/src/Ecal.cxx
@@ -1,76 +1,0 @@
-#include "DetDescr/Ecal.h"
-
-// LDMX
-#include "DetDescr/DetectorElement.h"
-#include "DetDescr/EcalDetectorID.h"
-
-// ROOT
-#include <Rtypes.h>
-#include <TGeoNode.h>
-#include <TNamed.h>
-#include <iomanip>
-#include <iostream>
-#include <sstream>
-#include <stdexcept>
-#include <string>
-
-namespace ldmx {
-
-    EcalStation::EcalStation(DetectorElementImpl* parent, TGeoNode* support) : DetectorElementImpl(parent, support) {
-
-        int copynum = support_->GetNumber();
-
-        // Follows convention where layer number is copynum divided by 7. --JMc
-        layerNum_ = copynum / 7;
-
-        // Follows convention where module number is copynum modulo 7. --JMc
-        moduleNum_ = copynum % 7;
-
-        // Ecal stations are named using their node copynum.
-        std::stringstream ss;
-        ss << std::setfill('0') << std::setw(3) << copynum;
-        name_ = "EcalStation" + ss.str();
-
-        // Create the bit-packed ID for the detector component.
-        auto detID = getDetectorID();
-        detID->setFieldValue(1, layerNum_);
-        detID->setFieldValue(2, moduleNum_);
-        id_ = getDetectorID()->pack();
-    }
-
-    Ecal::Ecal() {
-        name_ = "Ecal";
-        detectorID_ = new EcalDetectorID();
-    }
-
-    Ecal::~Ecal() {
-    }
-
-    void Ecal::initialize() {
-
-        if (!support_) {
-            throw std::runtime_error("The support is not set.");
-        }
-
-        if (!parent_) {
-            throw std::runtime_error("The parent is not set.");
-        }
-
-        detectorID_->clear();
-        detectorID_->setFieldValue(0, support_->GetNumber());
-        id_ = detectorID_->pack();
-
-        int nDau = this->support_->GetNdaughters();
-        TGeoNode* dau = nullptr;
-        for (int iDau = 0; iDau < nDau; iDau++) {
-            auto dau = support_->GetDaughter(iDau);
-            // Sensor names must start with "Si_" in the GDML!
-            static std::string prefix = "Si_";
-            if (!std::string(dau->GetName()).compare(0, prefix.size(), prefix)) {
-                new EcalStation(this, dau);
-            }
-        }
-    }
-
-    DE_ADD(Ecal)
-}
