+++ DetDescr/test/DefaultDetectorID_test.cxx
@@ -8,7 +8,6 @@
 
 using ldmx::IDField;
 using ldmx::DefaultDetectorID;
-using ldmx::FieldValueList;
 using ldmx::DetectorID;
 
 int main(int, const char* argv[])  {
@@ -29,7 +28,7 @@ int main(int, const char* argv[])  {
     std::cout << "rawValBits: " << rawValBits << std::endl;
 
     detID->setRawValue(rawVal);
-    const FieldValueList& fieldValues = detID->unpack();
+    const DetectorID::FieldValueList& fieldValues = detID->unpack();
 
     std::cout << "subdet from field values: " << fieldValues[0] << std::endl;
     std::cout << "layer from field values: " << fieldValues[1] << std::endl;
