+++ Framework/include/Framework/EventProcessor.h
@@ -12,7 +12,6 @@
 #include "Event/Event.h"
 #include "Event/RunHeader.h"
 #include "Framework/ParameterSet.h"
-#include "DetDescr/DetectorDataService.h"
 #include "Framework/StorageControl.h"
 
 // STL
@@ -67,14 +66,6 @@ namespace ldmx {
              * @param runHeader The RunHeader containing run information.
              */
             virtual void onNewRun(const RunHeader& runHeader) {
-            }
-
-            /**
-             * Callback for when there is new detector data activated.
-             * @param detectorService The DetectorDataService providing the detector data.
-             * @note This is called after the <i>onNewRun()</i> method.
-             */
-            virtual void onNewDetector(DetectorDataService* detectorService) {
             }
 
             /**
