+++ Framework/include/Framework/Process.h
@@ -9,11 +9,11 @@
 
 // LDMX
 #include "Framework/Exception.h"
-#include "DetDescr/DetectorDataServiceImpl.h"
 #include "Framework/StorageControl.h"
 
 // STL
 #include <vector>
+
 
 class TFile;
 class TDirectory;
@@ -161,9 +161,6 @@ namespace ldmx {
 
             /** TFile for histograms and other user products */
             TFile* histoTFile_{0};
-
-            /** Service providing access to detector data. */
-            DetectorDataServiceImpl* detectorService_{nullptr};
     };
 }
