+++ SimApplication/src/SimApplication.cxx
@@ -4,8 +4,6 @@
 #include "SimApplication/DetectorConstruction.h"
 #include "SimApplication/RunManager.h"
 #include "SimApplication/SimApplicationMessenger.h"
-#include "SimApplication/GDMLMessenger.h"
-#include "SimApplication/GDMLReadStructure.h"
 
 // STL
 #include <vector>
@@ -42,13 +40,8 @@ namespace ldmx {
         G4RunManager* runManager = new RunManager;
 
         // Setup GDML parser and messenger.
-        GDMLReadStructure* gdmlRead = new GDMLReadStructure;
-        G4GDMLParser* parser = new G4GDMLParser(gdmlRead);
-        parser->SetAddPointerToName(false);
-        parser->SetStripFlag(true);
-        G4UImessenger* g4gdmlMessenger = new G4GDMLMessenger(parser);
-        // Custom GDML messenger for exporting without pointers appended to names.
-        G4UImessenger* gdmlMessenger = new GDMLMessenger(parser, gdmlRead);
+        G4GDMLParser* parser = new G4GDMLParser();
+        G4UImessenger* gdmlMessenger = new G4GDMLMessenger(parser);
 
         // Create application messenger.
         new SimApplicationMessenger();
@@ -77,7 +70,6 @@ namespace ldmx {
             delete ui;
         }
 
-        delete g4gdmlMessenger;
         delete gdmlMessenger;
         delete parser;
         delete runManager;
